requirements.txt
orbitalis
paho-mqtt